pyproject.toml
[project]
name = "aaa-seg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "camhpj", email = "camhpjohnson@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "ipykernel>=6.29.5",
    "ipympl>=0.9.7",
    "ipython>=9.4.0",
    "ipywidgets>=8.1.7",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.0",
    "pynrrd>=1.1.3",
    "ruff>=0.12.2",
    "segmentation-models-pytorch>=0.5.0",
    "simpleitk>=2.5.2",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
